Cut and paste EffectUIServices to new files
