+++ src/effects/EffectUIServices.cpp
@@ -1,0 +1,12 @@
+/*!********************************************************************
+
+  Audacity: A Digital Audio Editor
+
+  @file EffectUIServices.cpp
+
+  Paul Licameli split from EffectPlugin.cpp
+
+**********************************************************************/
+#include "EffectUIServices.h"
+
+EffectUIServices::~EffectUIServices() = default;
